Refactor MentorSearch for readability
